# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfolio blog JSON API",
        "description": "Portfolio blog API documentation",
        "contact": {
            "email": "yseffk@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/blog/posts": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get posts  data whith pagination",
                "description": "Display a listing of the resource.",
                "operationId": "06f217e7801c10470d79c412475c3ba1",
                "parameters": [
                    {
                        "name": "pagination",
                        "in": "query",
                        "description": "Need pagination",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "is_published",
                        "in": "query",
                        "description": "filtering by 'is published' flag",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Per page limit (default 12)",
                        "required": false,
                        "example": "15"
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort  list criteria. Allowed fields are: id, sort, created_at. Prepend '-' sign to revert order",
                        "required": false,
                        "example": "-id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create post.",
                "operationId": "cc69d21b8ad8cff4671efdde22115371",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "group",
                                    "is_published",
                                    "sort"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Post title",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "is_published": {
                                        "description": "Post published flag",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "sort": {
                                        "description": "sort",
                                        "type": "integer",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Validation information.",
                                            "type": "string",
                                            "example": "The given data was invalid.",
                                            "nullable": false
                                        },
                                        "errors": {
                                            "description": "Object with all errors.",
                                            "type": "object",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/blog/posts/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Show post.",
                "description": "Display the specified resource.",
                "operationId": "51d8f99da8c1090fb2653507c17debf9",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "post id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/blog/posts/{post}?_method=PUT": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update Month block item.",
                "operationId": "1a35b0227354448e56284b3442d78cc1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "is_published",
                                    "sort"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Post title",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "is_published": {
                                        "description": "Post published flag",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "sort": {
                                        "description": "sort",
                                        "type": "integer",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Model not found"
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Validation information.",
                                            "type": "string",
                                            "example": "The given data was invalid.",
                                            "nullable": false
                                        },
                                        "errors": {
                                            "description": "Object with all errors.",
                                            "type": "object",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/blog/posts/{post}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete post ID",
                "operationId": "9e0078a6b01af62e8fc20e417298ecb5",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "post ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}